+++ src/discord/bot/analytics.ts
@@ -1,6 +1,12 @@
 import Discord from "discord.js";
+import moment from "moment";
+import { getAnalyticsErrorMessage } from "../../analytics/errors";
+import { PosthogEvent } from "../../analytics/events";
+import * as reports from "../../analytics/reports";
+import { fetchEventsForReportGenerator } from "../../analytics/reports";
+import { ChartReport, TextReport } from "../../analytics/reports/reports";
+import logger from "../../utils/logger";
 import { REPORTS_CHANNEL_ID } from "../channel-ids";
-import { sendAnalyticsReport } from "../reports";
 import { fetchTextChannelById } from "../utils";
 
 export function isAnalyticsMessage(message: Discord.Message): boolean {
@@ -102,4 +108,126 @@ If you want more control and generate some reports manually, you can check out
 wasp-lang/wasp-bot repo and generate them locally, README has instructions on this.
 `,
   );
+}
+
+export async function initiateAnalyticsReport(
+  discordClient: Discord.Client,
+): Promise<void> {
+  const reportsChannel = await fetchTextChannelById(
+    discordClient,
+    REPORTS_CHANNEL_ID,
+  );
+
+  await reportsChannel.send(
+    "📊 What time is it? It is time for daily analytics report!",
+  );
+  await reportsChannel.send("⏳ Fetching analytics events...");
+
+  try {
+    // By prefetching events, we can reuse them for all reports
+    const events = await fetchEventsForReportGenerator();
+
+    await sendAnalyticsReport(discordClient, "total", events);
+    await sendAnalyticsReport(discordClient, "daily", events);
+    if (isFirstDayOfWeek()) {
+      await sendAnalyticsReport(discordClient, "weekly", events);
+    }
+    if (isFirstDayOfMonth()) {
+      await sendAnalyticsReport(discordClient, "monthly", events);
+    }
+  } catch (e) {
+    logger.error(e);
+    const message = getAnalyticsErrorMessage(e);
+    await reportsChannel.send(
+      `Failed to send daily analytics report: ${message}`,
+    );
+  }
+}
+
+function isFirstDayOfWeek(): boolean {
+  return moment().isoWeekday() === 1;
+}
+
+function isFirstDayOfMonth(): boolean {
+  return moment().date() === 1;
+}
+
+type CliReportTypes = "daily" | "weekly" | "monthly" | "total";
+
+export async function sendAnalyticsReport(
+  discordClient: Discord.Client,
+  reportType: CliReportTypes,
+  prefetchedEvents: PosthogEvent[] | undefined = undefined,
+  numPeriods: number | undefined = undefined,
+): Promise<void> {
+  const waspReportsChannel = await fetchTextChannelById(
+    discordClient,
+    REPORTS_CHANNEL_ID,
+  );
+
+  waspReportsChannel.send(`⏳ Generating ${reportType} report...`);
+  const compositeReport = await getAnalyticsReport(
+    reportType,
+    prefetchedEvents,
+    numPeriods,
+  );
+
+  waspReportsChannel.send(
+    `=============== ${reportType.toUpperCase()} ANALYTICS REPORT ===============`,
+  );
+  for (const simpleReport of Object.values(compositeReport)) {
+    waspReportsChannel.send(convertSimpleReportToDiscordMessage(simpleReport));
+  }
+  waspReportsChannel.send(
+    "=======================================================",
+  );
+}
+
+function getAnalyticsReport(
+  reportType: CliReportTypes,
+  prefetchedEvents: PosthogEvent[] | undefined = undefined,
+  numPeriods: number | undefined = undefined,
+): Promise<{ [reportName: string]: Partial<TextReport & ChartReport> }> {
+  if (reportType === "monthly") {
+    return reports.generateMonthlyReport(prefetchedEvents, numPeriods);
+  } else if (reportType === "weekly") {
+    return reports.generateWeeklyReport(prefetchedEvents, numPeriods);
+  } else if (reportType === "daily") {
+    return reports.generateDailyReport(prefetchedEvents, numPeriods);
+  } else if (reportType === "total") {
+    return reports.generateTotalReport(prefetchedEvents);
+  } else {
+    throw new Error(`Unknown report type: ${reportType}`);
+  }
+}
+
+const DISCORD_MAX_MSG_SIZE = 2000;
+const DISCORD_MESSAGE_TOO_LONG_SUFFIX =
+  "\n... ⚠️ MESSAGE CUT BECAUSE IT IS TOO LONG...";
+
+function convertSimpleReportToDiscordMessage(
+  report: Partial<TextReport & ChartReport>,
+): Discord.MessageOptions {
+  const options: Discord.MessageOptions = {};
+  if (report.text) {
+    let content: string = report.text.join("\n");
+
+    if (content.length >= DISCORD_MAX_MSG_SIZE) {
+      content =
+        content.substring(
+          0,
+          DISCORD_MAX_MSG_SIZE - DISCORD_MESSAGE_TOO_LONG_SUFFIX.length,
+        ) + DISCORD_MESSAGE_TOO_LONG_SUFFIX;
+    }
+    options.content = content;
+  }
+
+  if (report.chart) {
+    const embed = new Discord.MessageEmbed();
+    embed.setImage(report.chart.toURL());
+
+    options.embed = embed;
+  }
+
+  return options;
 }
